+++ tcex/key_value_store/__init__.py
@@ -2,5 +2,6 @@
 
 # flake8:noqa
 from .key_value_api import KeyValueApi
+from .key_value_mock import KeyValueMock
 from .key_value_redis import KeyValueRedis
 from .redis_client import RedisClient