+++ tcex/tcex.py
@@ -21,7 +21,7 @@ from tcex.app_feature import AdvancedRequest
 from tcex.backports import cached_property
 from tcex.exit.exit import ExitCode, ExitService
 from tcex.input.input import Input
-from tcex.key_value_store import KeyValueApi, KeyValueRedis, RedisClient
+from tcex.key_value_store import KeyValueApi, KeyValueMock, KeyValueRedis, RedisClient
 from tcex.logger.logger import Logger  # pylint: disable=no-name-in-module
 from tcex.playbook import Playbook
 from tcex.pleb.proxies import proxies
@@ -290,6 +290,13 @@ class TcEx:
 
         if self.inputs.model_unresolved.tc_kvstore_type == 'TCKeyValueAPI':
             return KeyValueApi(self.session_tc)
+
+        if self.inputs.model_unresolved.tc_kvstore_type == 'Mock':
+            self.log.warning(
+                'Using mock key-value store.  '
+                'This should *never* happen when running in-platform.'
+            )
+            return KeyValueMock()
 
         raise RuntimeError(
             f'Invalid KV Store Type: ({self.inputs.model_unresolved.tc_kvstore_type})'